vh: sched: Better interaction between map_util_freq_pixel_mod() and uclamp
The 25% boost is unnecessary if the requested util is already boosted.

Or if we are capped by uclamp_max then we should not cross this
boundary.

Fix these two issues. We could do better in terms of code as there's
duplication of effort, but refactoring will be more expensive FWIW.

Bug: 270955580
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: If197d0e1252a326b29d02c870043bc7d3dea3886
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>