vh: sched: Pass the cpu with max util
In get_next_freq, it will pass policy->cpu to map_util_freq_pixel_mod,
but in the recent change of map_util_freq_pixel_mod, it will get cpu
util of the cpu passed in. It will be policy->cpu which is wrong. We
will need to pass the cpu with the max util.

Bug: 277860694
Bug: 230707523
Change-Id: Ia1c627aab930c1504d4ea61b7b62dc6d7d8f8e19
Signed-off-by: Rick Yiu <rickyiu@google.com>
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>