vh: sched: Fix prev_cpu check in feec()
The prev_cpu check in feec() incorrectly checks if prev_cpu is present in
the allowed mask, though it may not be present in the candidates mask. If
it's part of the allowed mask but not the candidates mask, then prev_cpu
won't have any statistics computed for it and thus the prev_cpu check can
incorrectly return prev_cpu even though prev_cpu isn't actually equivalent
to best_cpu.

Fix it by checking if prev_cpu is actually part of the candidates mask
instead. Since best_cpu is always removed from the candidates mask by the
time prev_cpu is checked, this also means that prev_cpu's presence in
the candidates mask implies prev_cpu != best_cpu already.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>